Made Allegro more 64-bit friendly: - include inttypes.h/stdint.h or otherwise compensate if not available; - use uintptr_t to hold memory addresses instead of unsigned long; - made 32-bit bitmaps really be arrays of 32-bit integers instead of   arrays of unsigned longs (also made 16-bit bitmaps be arrays of 16-bit   integers instead of unsigned shorts).

+++ examples/exrgbhsv.c
@@ -137,7 +137,7 @@ int update_color_value(void *dp3, int val)
 {
    /* 'val' is the value of the slider's position (0-255),
       'type' is which slider was changed */
-   int type = ((unsigned long)dp3 - (unsigned long)colors) / sizeof(colors[0]);
+   int type = ((uintptr_t)dp3 - (uintptr_t)colors) / sizeof(colors[0]);
    int r, g, b;
    float h, s, v;
