+++ include/allegro/platform/al386vc.h
@@ -44,7 +44,7 @@ END_OF_INLINE(_default_ds);
 /* bmp_write_line:
  *  Bank switch function.
  */
-INLINE _AL_DLL unsigned long bmp_write_line(BITMAP *bmp, int line)
+INLINE _AL_DLL uintptr_t bmp_write_line(BITMAP *bmp, int line)
 { 
    _asm { 
       mov edx, bmp
@@ -61,7 +61,7 @@ END_OF_INLINE(bmp_write_line);
 /* bmp_read_line:
  *  Bank switch function.
  */
-INLINE _AL_DLL unsigned long bmp_read_line(BITMAP *bmp, int line)
+INLINE _AL_DLL uintptr_t bmp_read_line(BITMAP *bmp, int line)
 {
    _asm {
       mov edx, bmp
