+++ src/unix/jack.c
@@ -135,7 +135,7 @@ static int jack_process (jack_nframes_t nframes, void *arg)
    unsigned short *buffer16 = jack_buffer;
    unsigned char *buffer8 = jack_buffer;
 
-   _mix_some_samples((unsigned long) jack_buffer, 0, jack_signed);
+   _mix_some_samples((uintptr_t) jack_buffer, 0, jack_signed);
 
    jack_default_audio_sample_t *out_left = (jack_default_audio_sample_t *)
       jack_port_get_buffer (output_left, nframes);
@@ -253,7 +253,7 @@ static int jack_init(int input, int voices)
       return -1;
    }
 
-   _mix_some_samples((unsigned long) jack_buffer, 0, jack_signed);
+   _mix_some_samples((uintptr_t) jack_buffer, 0, jack_signed);
 
    if (jack_activate (jack_client)) {
       ustrzcpy(allegro_error, ALLEGRO_ERROR_SIZE, get_config_text(
