+++ src/unix/sgial.c
@@ -142,7 +142,7 @@ static void *_al_sgial_puller_thread_func(void *arg)
       FD_SET(fd, &fds);
       select(FD_SETSIZE, NULL, &fds, NULL, NULL);
       alWriteFrames(_al_sgial_port, _al_sgial_bufdata, _al_sgial_bufsize);
-      _mix_some_samples((unsigned long) _al_sgial_bufdata, 0, _al_sgial_signed);
+      _mix_some_samples((uintptr_t) _al_sgial_bufdata, 0, _al_sgial_signed);
    }
 }
 
@@ -155,7 +155,7 @@ static void _al_sgial_update(int threaded)
 {
    if (alGetFillable(_al_sgial_port) > _al_sgial_bufsize) {
       alWriteFrames(_al_sgial_port, _al_sgial_bufdata, _al_sgial_bufsize);
-      _mix_some_samples((unsigned long) _al_sgial_bufdata, 0, _al_sgial_signed);
+      _mix_some_samples((uintptr_t) _al_sgial_bufdata, 0, _al_sgial_signed);
    }
 }
 
@@ -252,7 +252,7 @@ static int _al_sgial_init(int input, int voices)
       return -1;
    }
 
-   _mix_some_samples((unsigned long) _al_sgial_bufdata, 0, _al_sgial_signed);
+   _mix_some_samples((uintptr_t) _al_sgial_bufdata, 0, _al_sgial_signed);
 
 #ifdef HAVE_LIBPTHREAD
    /* Add audio thread. */
